% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_chatgpt.R
\name{ask_chatgpt}
\alias{ask_chatgpt}
\title{Ask ChatGPT}
\usage{
ask_chatgpt(
  question,
  knowledge_dir = NULL,
  context = NULL,
  history_file = NULL,
  update = FALSE
)
}
\arguments{
\item{question}{The question to ask ChatGPT.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
testing
}
\examples{
\dontrun{
cat(ask_chatgpt("What do you think about R language?"))
}

}
