% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_chatgpt.R
\name{ask_chatgpt_pe}
\alias{ask_chatgpt_pe}
\title{Performing prompt engineering tasks}
\usage{
ask_chatgpt_pe(q, pe_list, add_req = "")
}
\arguments{
\item{q}{question}

\item{pe_list}{list of prompt_engineering_functions}

\item{add_req}{additional requests}
}
\description{
Performing prompt engineering tasks
}
