% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_examples.R
\name{chat_apa_rename}
\alias{chat_apa_rename}
\title{Rename Article File Name}
\usage{
chat_apa_rename(citation)
}
\arguments{
\item{citation}{apa6th citation}
}
\description{
Rename Article File Name
}
