% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_examples.R
\name{chat_citation_end}
\alias{chat_citation_end}
\title{Get an English citation Sentence}
\usage{
chat_citation_end(authors, year, q)
}
\arguments{
\item{authors}{authors name}

\item{year}{publication year}

\item{q}{query}
}
\description{
Get an English citation Sentence
}
