% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_examples.R
\name{chat_edit_writing}
\alias{chat_edit_writing}
\title{Edit writing}
\usage{
chat_edit_writing(q)
}
\arguments{
\item{q}{writing content}
}
\description{
Provides the ability to correct English sentences
}
