% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_examples.R
\name{chat_get_keywords_textrank}
\alias{chat_get_keywords_textrank}
\title{Extracting Keywords - textrank}
\usage{
chat_get_keywords_textrank(q)
}
\arguments{
\item{q}{sentence}
}
\description{
Extracting Keywords
}
