% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_examples.R
\name{chat_news_head}
\alias{chat_news_head}
\title{Get News Title Using Google Search Engine}
\usage{
chat_news_head(q, kor = TRUE)
}
\arguments{
\item{q}{news query}

\item{kor}{if TRUE, output is in Korean, o.w., English}
}
\description{
Get News Title Using Google Search Engine
}
