% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_examples.R
\name{chat_script}
\alias{chat_script}
\title{Script Writer}
\usage{
chat_script(q, sec = 30)
}
\arguments{
\item{q}{keywords}

\item{sec}{length measured in seconds (default = 30 seconds)}
}
\description{
Writing a sentence based on keywords supplied.
}
