% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_examples.R
\name{chat_topic_check}
\alias{chat_topic_check}
\title{Topic checker}
\usage{
chat_topic_check(q, topic_keyword)
}
\arguments{
\item{q}{Sentence to check}

\item{topic_keyword}{string vector, topics}
}
\description{
Checking topic and produce results
}
