% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_list.R
\name{compile_prompt}
\alias{compile_prompt}
\title{Compile prompt engineering functions}
\usage{
compile_prompt(pe_list)
}
\arguments{
\item{pe_list}{list of prompt engineering functions}
}
\description{
Compile prompt engineering functions
}
