% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_list.R
\name{create_knowledge_repository}
\alias{create_knowledge_repository}
\title{Create Session History}
\usage{
create_knowledge_repository(tmp_file)
}
\arguments{
\item{tmp_file}{temporar file}
}
\description{
Create session history
}
\examples{
sesison <- tempfile()
create_knowledge_repository(session)
ask_chatgpt("Remember following information.",
  history_file = session,
  update = TRUE)
ask_chatgpt("I like an apple.",
            history_file = session,
            update = TRUE)
ask_chatgpt("And, I like an orange.",
            history_file = session,
            update = TRUE)
ask_chatgpt("What are my favorite fruit?",
            history_file = session,
            update = TRUE)
file.remove(session)
}
