% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_writer.R
\name{create_writer_xls}
\alias{create_writer_xls}
\title{Create a template file for writing English papers}
\usage{
create_writer_xls(fname)
}
\description{
Create a template file for writing English papers
}
