% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webresource.R
\name{get_gs_news_title}
\alias{get_gs_news_title}
\title{Get News Title from Google Search}
\usage{
get_gs_news_title(search_query)
}
\arguments{
\item{search_query}{What to search for}
}
\value{
Combine search results into a single sentence
}
\description{
Get the first page results of a news search in Google Search.
}
