% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webresource.R
\name{get_naver_news}
\alias{get_naver_news}
\title{Get News Title from Naver Search}
\usage{
get_naver_news(search_query, client_id, client_secret)
}
\arguments{
\item{search_query}{What to search for}

\item{client_id}{Client ID}

\item{client_secret}{Client secret}
}
\description{
Get the first page results of a news search in Naver Search API.
}
