% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering.R
\name{pe_citation_end}
\alias{pe_citation_end}
\title{Prompt Engineering Function - Citation at the end of a sentence}
\usage{
pe_citation_end()
}
\description{
Prompt Engineering Function - Citation at the end of a sentence
}
