% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering_writer.R
\name{pe_edit_eng_writing}
\alias{pe_edit_eng_writing}
\title{Prompt Enginnering Function - Editing a paragraph in academic styling}
\usage{
pe_edit_eng_writing()
}
\description{
Prompt Enginnering Function - Editing a paragraph in academic styling
}
