% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering.R
\name{pe_edit_kor_writing}
\alias{pe_edit_kor_writing}
\title{Prompt Enginnering Function - Editing a paragraph in academic styling (Korean)}
\usage{
pe_edit_kor_writing(knowledge = NULL)
}
\description{
Editing a paragraph in academic styling (Korean)
}
