% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering.R
\name{pe_eng_script_writing}
\alias{pe_eng_script_writing}
\title{Prompt Engineering Function - Generate a script for Powerpoint presentation}
\usage{
pe_eng_script_writing(knowledge = NULL, seconds = 20)
}
\description{
Generate a script for Powerpoint presentation
}
