% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering.R
\name{pe_eng_title}
\alias{pe_eng_title}
\title{Prompt Engineering Function - Generating a news title in English}
\usage{
pe_eng_title(knowledge = NULL)
}
\description{
Prompt Engineering Function - Generating a news title in English
}
