% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering.R
\name{pe_get_topic_sentece}
\alias{pe_get_topic_sentece}
\title{Prompt Engineering - Generating a Topic Sentence}
\usage{
pe_get_topic_sentece(knowledge = NULL)
}
\description{
Prompt Engineering - Generating a Topic Sentence
}
