% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering.R
\name{pe_news_title}
\alias{pe_news_title}
\title{Prompt Engineering Function - Reporting news summary}
\usage{
pe_news_title(search_query, naver_client_id = NULL, naver_client_secret = NULL)
}
\description{
Prompt Engineering Function - Reporting news summary
}
