% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering.R
\name{pe_write_eng_sentence_based_idea}
\alias{pe_write_eng_sentence_based_idea}
\title{Prompt Enginnering Function - Sentence Writer Based on Supplied Keywords}
\usage{
pe_write_eng_sentence_based_idea(knowledge = NULL)
}
\description{
Sentence Writer Based on Supplied Keywords
}
