% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_chatgpt.R
\name{teach_chatgpt}
\alias{teach_chatgpt}
\title{Teach ChatGPT}
\usage{
teach_chatgpt(question, file = NULL)
}
\arguments{
\item{question}{Teach ChatGPT, given JSON data}

\item{file}{JSON data}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
Teach ChatGPT
}
\examples{
\dontrun{
cat(ask_chatgpt("What do you think about R language?"))
}

}
