% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_list.R
\name{teach_gpt}
\alias{teach_gpt}
\title{Teach GPT with session}
\usage{
teach_gpt(tmp_file, intent = NULL, me = NULL, you = NULL)
}
\arguments{
\item{tmp_file}{session temporary file}

\item{intent}{(option) context information}

\item{me}{(option) your quesiton}

\item{you}{(option) answer by GPT}
}
\description{
Teach GPT with session
}
