% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_list.R
\name{transform_excel_data_to_json}
\alias{transform_excel_data_to_json}
\title{Transform Excel conversation data to JSON data}
\usage{
transform_excel_data_to_json(input_exel, output_json)
}
\arguments{
\item{output_json}{Output JSON file with an extension (i.e., .json)}

\item{input_excel}{Excel file with an extension (i.e., .xlsx)}
}
\description{
Transform Excel conversation data to JSON data
}
