% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_list.R
\name{create_session_history}
\alias{create_session_history}
\title{Create Session History}
\usage{
create_session_history(tmp_file)
}
\arguments{
\item{tmp_file}{temporar file}
}
\description{
Create session history
}
\examples{
sesison <- tempfile()
create_session_history(session)
ask_chatgpt("Hello!",history_file=session)
# if you want to delete the session temp file, run the following:
file.remove(session)
}
