% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering.R
\name{pe_eng_script_writing}
\alias{pe_eng_script_writing}
\title{Prompt Enginnering Function - Generate a script for Powerpoint presentation (30 seconds)}
\usage{
pe_eng_script_writing(seconds = 30)
}
\description{
Generate a script for Powerpoint presentation (30 seconds)
}
