% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering_writer.R
\name{pe_topic_flag}
\alias{pe_topic_flag}
\title{Prompt Enginnering Function - Checking the inclusion of keywords}
\usage{
pe_topic_flag()
}
\description{
Prompt Enginnering Function - Checking the inclusion of keywords
}
